+++ nodejs/deleteHandler.js
@@ -38,10 +38,12 @@ class DeleteHandler {
      *
      * @param {Object} declaration - Parsed declaration of objects to delete.
      * @param {Object} bigIp - BigIp object.
+     * @param {EventEmitter} - DO event emitter.
      */
-    constructor(declaration, bigIp) {
+    constructor(declaration, bigIp, eventEmitter) {
         this.declaration = declaration;
         this.bigIp = bigIp;
+        this.eventEmitter = eventEmitter;
     }
 
     /**